+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 0.0.6
+- WIN: added Chromium Embedded Framework version
+- Utilised observer model better
+- ON_PROPERTY_CHANGE event added
+- Removed unneeded events
+- updated example
+
 ### 0.0.5
 - back forward methods added